veel veranderingen

+++ KBS2k19/src/DBServer.java
@@ -1,5 +1,5 @@
 public class DBServer extends Component {
-    public DBServer(String naam, int beschikbaarheid, int kosten) {
+    public DBServer(String naam, double beschikbaarheid, int kosten) {
         super(naam, "DBServer", beschikbaarheid, kosten);
     }
 }
